Flask/requirements.txt
# For Web Core Framework
Flask==2.3.3
Werkzeug==2.3.7 

# Database
firebase-admin==6.4.0
google-cloud-firestore

# Ai Api
openai==1.6.1

# Security
python-dotenv==1.0.0
flask-limiter==3.5.0
flask-talisman==1.1.0
limits==3.6.0

# Ai & Ml
Pillow==10.0.0
numpy==1.24.3